requirements.txt
jieba
chinese_converter
synonyms
jieba.posseg
csv
matplotlib.pyplot
pandas
spacy